DataMining/Classification/requirements.txt
pandas==0.25.0
numpy==1.17.0
matplotlib==3.1.1
python_igraph==0.7.1.post6
